feat: add bucket list API endpoints with model controller and routes
