--- back-end/.env.template
@@ -1,1 +1,0 @@
-PORT =
